refactor: move IConnectToWalletOptions to shared to fix kit-bg → kit import violation
