+++ packages/shared/types/externalWallet.types.ts
@@ -1,3 +1,5 @@
+import type { IKeylessPendingLogin } from '../src/keylessWallet/keylessWebTypes';
+
 import type { ExternalConnectorWalletConnect } from '@onekeyhq/kit-bg/src/connectors/chains/walletconnect/ExternalConnectorWalletConnect';
 import type { IDBAccountAddressesMap } from '@onekeyhq/kit-bg/src/dbs/local/types';
 import type { WalletConnectDappSideProvider } from '@onekeyhq/kit-bg/src/services/ServiceWalletConnect/WalletConnectDappSideProvider';
@@ -200,4 +202,11 @@ export type IExternalConnectWalletResult = {
   connectionInfo: IExternalConnectionInfo;
   accountInfo: IExternalConnectAccountInfo;
   notSupportedNetworkIds: string[] | undefined;
+};
+
+export type IConnectToWalletOptions = {
+  allowEmptyAuthorizedAddresses?: boolean;
+  suppressDeniedToast?: boolean;
+  skipDisconnectConnector?: boolean;
+  webKeylessPendingLogin?: IKeylessPendingLogin;
 };
